@@ -2,10 +2,10 @@
 #define EP_BITMAPFONT_H
 
 #include "bitmapfont_glyph.h"
-#include "shinonome_gothic.h"
-#include "shinonome_mincho.h"
-#include "bitmapfont_rmg2000.h"
-#include "bitmapfont_ttyp0.h"
-#include "bitmapfont_wqy.h"
+#include "generated/shinonome_gothic.h"
+#include "generated/shinonome_mincho.h"
+#include "generated/bitmapfont_rmg2000.h"
+#include "generated/bitmapfont_ttyp0.h"
+#include "generated/bitmapfont_wqy.h"
 
 #endif // EP_BITMAPFONT_H
