@@ -24,9 +24,12 @@
 // Headers
 #include "audio_decoder.h"
 #define DR_WAV_NO_STDIO
-#include "external/dr_wav.h"
-#include <string>
-#include <memory>
+// Use system dr_wav header if available
+#if __has_include(<dr_wav.h>)
+#  include <dr_wav.h>
+#else
+#  include "external/dr_wav.h"
+#endif
 
 /**
  * Standalone audio decoder powered by dr_wav
