@@ -46,10 +46,13 @@ public:
 		uint64_t version = 0;
 		int logos = 0;
 		std::string version_str;
+		uint32_t code_size = 0;
+		uint32_t cherry_size = 0;
+		uint32_t geep_size = 0;
 		bool is_i386 = true;
 		bool is_easyrpg_player = false;
 
-		int GetEngineType() const;
+		int GetEngineType(bool& is_maniac_patch) const;
 		void Print() const;
 	};
 
