@@ -16,14 +16,16 @@
  */
 
  // Headers
-#include <lcf/data.h>
-#include "compiler.h"
-#include "string_view.h"
 #include <cstdint>
 #include <string>
-#include "output.h"
-#include "main_data.h"
+#include <lcf/data.h>
+#include "compiler.h"
 #include "game_variables.h"
+#include "main_data.h"
+#include "output.h"
+#include "pending_message.h"
+#include "player.h"
+#include "string_view.h"
 
 /**
  * Game_Strings class.
@@ -35,6 +37,10 @@ public:
 
 	static constexpr int max_warnings = 10;
 
+	struct Str_Params {
+		int string_id, hex, extract;
+	};
+
 	Game_Strings();
 
 	void SetData(Strings_t s);
@@ -42,17 +48,44 @@ public:
 
 	Str_t Get(int id) const;
 	Str_t GetIndirect(int id) const;
+	Str_t GetWithMode(Str_t str_data, int mode, int arg) const;
+	Str_t GetWithModeAndPos(Str_t str_data, int mode, int arg, int* pos);
 
-	Str_t Asg(int string_id, Str_t string);
-	Str_t Cat(int string_id, Str_t string);
+	Str_t Asg(Str_Params params, Str_t string);
+	Str_t Cat(Str_Params params, Str_t string);
+	int ToNum(Str_Params params, int var_id);
+	int GetLen(Str_Params params, int var_id);
+	int InStr(Str_Params params, std::string search, int var_id, int begin = 0);
+	int Split(Str_Params params, std::string delimiter, int string_out_id, int var_id);
+	Str_t PopLine(Str_Params params, int offset, int string_out_id);
+	Str_t ExMatch(Str_Params params, std::string expr, int var_id, int begin, int string_out_id = -1);
+
+	const Strings_t& RangeOp(Str_Params params, int string_id_1, Str_t string, int op, int args[] = nullptr);
+
+	Str_t PrependMin(Str_t string, int min_size, char c);
 
 private:
+	Str_t Set(Str_Params params, Str_t string);
+	bool ResizeWithId(int id);
 	bool ShouldWarn(int id) const;
 	void WarnGet(int id) const;
+	Str_t Extract(Str_t string);
+
 private:
 	Strings_t _strings;
 	mutable int _warnings = max_warnings;
 };
+
+
+inline Game_Strings::Str_t Game_Strings::Set(Str_Params params, Str_t string) {
+	if (!ResizeWithId(params.string_id)) return "";
+
+	auto& s = _strings[params.string_id - 1];
+	s = string;
+	if (params.extract)
+		s = Extract(s);
+	return s;
+}
 
 inline void Game_Strings::SetData(Strings_t s) {
 	_strings = std::move(s);
@@ -76,8 +109,49 @@ inline Game_Strings::Str_t Game_Strings::Get(int id) const {
 	return _strings[id - 1];
 }
 
+inline bool Game_Strings::ResizeWithId(int id) {
+	if (EP_UNLIKELY(ShouldWarn(id))) {
+		WarnGet(id);
+	}
+	if (id <= 0) {
+		return false;
+	}
+	if (EP_UNLIKELY(id > static_cast<int>(_strings.size()))) {
+		_strings.resize(id, "");
+	}
+	return true;
+}
+
 inline Game_Strings::Str_t Game_Strings::GetIndirect(int id) const {
 	auto val_indirect = Main_Data::game_variables->Get(id);
 	return Get(static_cast<int>(val_indirect));
 }
 
+inline Game_Strings::Str_t Game_Strings::GetWithMode(Str_t str_data, int arg, int mode) const {
+	switch (mode) {
+	case 1: // direct string reference
+		return Get(arg);
+		break;
+	case 2: // indirect string reference
+		return GetIndirect(arg);
+		break;
+	}
+	return str_data;
+}
+
+inline Game_Strings::Str_t Game_Strings::GetWithModeAndPos(Str_t str_data, int arg, int mode, int* pos) {
+	Str_t ret;
+	switch (mode) {
+	case 0:
+		ret = static_cast<Str_t>(static_cast<std::string>(str_data).substr(*pos, arg));
+		*pos += arg;
+		break;
+	case 1: // direct string reference
+		ret = Get(arg);
+		break;
+	case 2: // indirect string reference
+		ret = GetIndirect(arg);
+		break;
+	}
+	return ret;
+}
