@@ -64,10 +64,11 @@ class PendingMessage {
 
 		void SetIsEventMessage(bool value) { is_event_message = value; }
 		bool IsEventMessage() const { return is_event_message; }
+
+		static std::string ApplyTextInsertingCommands(std::string input, uint32_t escape_char, bool maniacs_parsing = false);
+
 	private:
 		int PushLineImpl(std::string msg);
-
-		std::string ApplyTextInsertingCommands(std::string input, uint32_t escape_char);
 
 	private:
 		ChoiceContinuation choice_continuation;
